fix(maintenance): fix extend_protocol_ttl compilation, event emission, tests and docs

+++ quicklendx-contracts/src/notifications.rs
@@ -3,11 +3,11 @@ use crate::protocol_limits::{
 };
 use crate::types::Bid;
 use crate::types::{Invoice, InvoiceStatus};
-use soroban_sdk::{contracttype, symbol_short, Address, Bytes, BytesN, Env, Map, String, Vec};
+use soroban_sdk::{contracttype, symbol_short, xdr::ToXdr, Address, Bytes, BytesN, Env, Map, String, Vec};
 
 /// Maximum number of idempotency keys to track in the bloom-resistant set.
 /// This provides protection against replay attacks while maintaining reasonable storage.
-const MAX_IDEMPOTENCY_KEYS: usize = 10_000;
+const MAX_IDEMPOTENCY_KEYS: u32 = 10_000;
 
 /// Notification types for different events
 #[contracttype]
